Implemented LSR

+++ CPU.h
@@ -99,7 +99,7 @@ class CPU {
     void LDX(uint8_t);
     void LDY(uint8_t);
     void LSR();
-    void LSR(uint8_t);
+    void LSR(uint16_t);
     void NOP();
     void ORA(uint8_t);
     void PHA();
